+++ src/main/java/org/asamk/signal/commands/SendCommand.java
@@ -12,6 +12,7 @@ import org.asamk.signal.manager.api.AttachmentInvalidException;
 import org.asamk.signal.manager.api.InvalidStickerException;
 import org.asamk.signal.manager.api.Message;
 import org.asamk.signal.manager.api.RecipientIdentifier;
+import org.asamk.signal.manager.api.TextStyle;
 import org.asamk.signal.manager.api.UnregisteredRecipientException;
 import org.asamk.signal.manager.groups.GroupNotFoundException;
 import org.asamk.signal.manager.groups.GroupSendingNotAllowedException;
@@ -66,6 +67,9 @@ public class SendCommand implements JsonRpcLocalCommand {
         subparser.addArgument("--mention")
                 .nargs("*")
                 .help("Mention another group member (syntax: start:length:recipientNumber)");
+        subparser.addArgument("--text-style")
+                .nargs("*")
+                .help("Style parts of the message text (syntax: start:length:STYLE)");
         subparser.addArgument("--quote-timestamp")
                 .type(long.class)
                 .help("Specify the timestamp of a previous message with the recipient or group to add a quote to the new message.");
@@ -74,6 +78,9 @@ public class SendCommand implements JsonRpcLocalCommand {
         subparser.addArgument("--quote-mention")
                 .nargs("*")
                 .help("Quote with mention of another group member (syntax: start:length:recipientNumber)");
+        subparser.addArgument("--quote-text-style")
+                .nargs("*")
+                .help("Quote with style parts of the message text (syntax: start:length:STYLE)");
         subparser.addArgument("--sticker").help("Send a sticker (syntax: stickerPackId:stickerId)");
         subparser.addArgument("--preview-url")
                 .help("Specify the url for the link preview (the same url must also appear in the message body).");
@@ -146,6 +153,9 @@ public class SendCommand implements JsonRpcLocalCommand {
         List<String> mentionStrings = ns.getList("mention");
         final var mentions = mentionStrings == null ? List.<Message.Mention>of() : parseMentions(m, mentionStrings);
 
+        List<String> textStyleStrings = ns.getList("text-style");
+        final var textStyles = textStyleStrings == null ? List.<TextStyle>of() : parseTextStyles(textStyleStrings);
+
         final Message.Quote quote;
         final var quoteTimestamp = ns.getLong("quote-timestamp");
         if (quoteTimestamp != null) {
@@ -155,10 +165,15 @@ public class SendCommand implements JsonRpcLocalCommand {
             final var quoteMentions = quoteMentionStrings == null
                     ? List.<Message.Mention>of()
                     : parseMentions(m, quoteMentionStrings);
+            List<String> quoteTextStyleStrings = ns.getList("quote-text-style");
+            final var quoteTextStyles = quoteTextStyleStrings == null
+                    ? List.<TextStyle>of()
+                    : parseTextStyles(quoteTextStyleStrings);
             quote = new Message.Quote(quoteTimestamp,
                     CommandUtil.getSingleRecipientIdentifier(quoteAuthor, m.getSelfNumber()),
                     quoteMessage == null ? "" : quoteMessage,
-                    quoteMentions);
+                    quoteMentions,
+                    quoteTextStyles);
         } else {
             quote = null;
         }
@@ -201,7 +216,8 @@ public class SendCommand implements JsonRpcLocalCommand {
                     Optional.ofNullable(quote),
                     Optional.ofNullable(sticker),
                     previews,
-                    Optional.ofNullable((storyReply)));
+                    Optional.ofNullable((storyReply)),
+                    textStyles);
             var results = editTimestamp != null
                     ? m.sendEditMessage(message, recipientIdentifiers, editTimestamp)
                     : m.sendMessage(message, recipientIdentifiers);
@@ -235,6 +251,30 @@ public class SendCommand implements JsonRpcLocalCommand {
                     m.getSelfNumber()), Integer.parseInt(matcher.group(1)), Integer.parseInt(matcher.group(2))));
         }
         return mentions;
+    }
+
+    private List<TextStyle> parseTextStyles(
+            final List<String> textStylesStrings
+    ) throws UserErrorException {
+        List<TextStyle> textStyles;
+        final Pattern textStylePattern = Pattern.compile("(\\d+):(\\d+):(.+)");
+        textStyles = new ArrayList<>();
+        for (final var textStyle : textStylesStrings) {
+            final var matcher = textStylePattern.matcher(textStyle);
+            if (!matcher.matches()) {
+                throw new UserErrorException("Invalid textStyle syntax ("
+                        + textStyle
+                        + ") expected 'start:length:STYLE'");
+            }
+            final var style = TextStyle.Style.from(matcher.group(3));
+            if (style == null) {
+                throw new UserErrorException("Invalid style: " + matcher.group(3));
+            }
+            textStyles.add(new TextStyle(style,
+                    Integer.parseInt(matcher.group(1)),
+                    Integer.parseInt(matcher.group(2))));
+        }
+        return textStyles;
     }
 
     private Message.Sticker parseSticker(final String stickerString) throws UserErrorException {