+++ src/main/java/org/asamk/signal/dbus/DbusSignalImpl.java
@@ -219,7 +219,8 @@ public class DbusSignalImpl implements Signal {
                             Optional.empty(),
                             Optional.empty(),
                             List.of(),
-                            Optional.empty()),
+                            Optional.empty(),
+                            List.of()),
                     getSingleRecipientIdentifiers(recipients, m.getSelfNumber()).stream()
                             .map(RecipientIdentifier.class::cast)
                             .collect(Collectors.toSet()));
@@ -388,7 +389,8 @@ public class DbusSignalImpl implements Signal {
                     Optional.empty(),
                     Optional.empty(),
                     List.of(),
-                    Optional.empty()), Set.of(RecipientIdentifier.NoteToSelf.INSTANCE));
+                    Optional.empty(),
+                    List.of()), Set.of(RecipientIdentifier.NoteToSelf.INSTANCE));
             checkSendMessageResults(results);
             return results.timestamp();
         } catch (AttachmentInvalidException e) {
@@ -431,7 +433,8 @@ public class DbusSignalImpl implements Signal {
                     Optional.empty(),
                     Optional.empty(),
                     List.of(),
-                    Optional.empty()), Set.of(getGroupRecipientIdentifier(groupId)));
+                    Optional.empty(),
+                    List.of()), Set.of(getGroupRecipientIdentifier(groupId)));
             checkSendMessageResults(results);
             return results.timestamp();
         } catch (IOException | InvalidStickerException e) {